Moved command line argument handling to main.rs.
